requirements.txt
openai
pandas
yfinance
matplotlib
python-dotenv
langgraph
langchain
langchain_openai